requirements.txt
pyttsx3
speedtest-cli>=2.1.2
matplotlib>=2.2.3